MODTLR-24 refactoring
